refactor remote explorer, wsl worked
