+++ src/himena/workflow/_graph.py
@@ -9,6 +9,7 @@ from himena.workflow._reader import (
     ProgrammaticMethod,
     LocalReaderMethod,
     RemoteReaderMethod,
+    WslReaderMethod,
     UserInput,
 )
 from himena.workflow._command import CommandExecution, UserModification
@@ -23,6 +24,7 @@ WorkflowStepType = Union[
     ProgrammaticMethod,
     LocalReaderMethod,
     RemoteReaderMethod,
+    WslReaderMethod,
     CommandExecution,
     UserModification,
     UserInput,
