+++ src/himena/workflow/_reader.py
@@ -67,6 +67,23 @@ class ReaderMethod(NoParentWorkflow):
         return store.run(path, plugin=self.plugin)
 
 
+class PathReaderMethod(ReaderMethod):
+    """Describes that one was read from a path."""
+
+    path: Path | list[Path]
+    force_directory: bool = Field(default=False)
+
+    def run_command(self, dst_path: Path, stdout=None):
+        """Run command to move the file from source to local `dst_path`."""
+        raise NotImplementedError("No conversion command defined for this method.")
+
+    def to_str(self) -> str:
+        """Return the local file path representation."""
+        if isinstance(self.path, Path):
+            return self.path.as_posix()
+        return ";".join(p.as_posix() for p in self.path)
+
+
 class LocalReaderMethod(ReaderMethod):
     """Describes that one was read from a local source file."""
 
@@ -98,17 +115,15 @@ class LocalReaderMethod(ReaderMethod):
         return model
 
 
-class RemoteReaderMethod(ReaderMethod):
+class RemoteReaderMethod(PathReaderMethod):
     """Describes that one was read from a remote source file."""
 
     type: Literal["remote-reader"] = "remote-reader"
     host: str
     username: str
-    path: Path
     port: int = Field(default=22)
     wsl: bool = Field(default=False)
     protocol: str = Field(default="rsync")
-    force_directory: bool = Field(default=False)
 
     @classmethod
     def from_str(
@@ -122,10 +137,14 @@ class RemoteReaderMethod(ReaderMethod):
     ) -> "RemoteReaderMethod":
         username, rest = s.split("@")
         host, path = rest.split(":")
+        if ";" in path:
+            path = [Path(p) for p in path.split(";")]
+        else:
+            path = Path(path)
         return cls(
             username=username,
             host=host,
-            path=Path(path),
+            path=path,
             wsl=wsl,
             protocol=protocol,
             output_model_type=output_model_type,
@@ -134,7 +153,11 @@ class RemoteReaderMethod(ReaderMethod):
 
     def to_str(self) -> str:
         """Return the remote file path representation."""
-        return f"{self.username}@{self.host}:{self.path.as_posix()}"
+        if isinstance(self.path, Path):
+            return f"{self.username}@{self.host}:{self.path.as_posix()}"
+        return (
+            f"{self.username}@{self.host}:{';'.join(p.as_posix() for p in self.path)}"
+        )
 
     def _get_model_impl(self, wf: "Workflow") -> "WidgetDataModel":
         out = self.run()
@@ -145,8 +168,15 @@ class RemoteReaderMethod(ReaderMethod):
 
     def run(self):
         with tempfile.TemporaryDirectory() as tmpdir:
-            dst_path = Path(tmpdir).joinpath(self.path.name)
-            self.run_command(dst_path)
+            if isinstance(self.path, Path):
+                dst_path = Path(tmpdir).joinpath(self.path.name)
+                self.run_command(dst_path)
+            else:
+                dst_path: list[Path] = []
+                for p in self.path:
+                    dst_p = Path(tmpdir).joinpath(p.name)
+                    self.run_command(dst_p)
+                    dst_path.append(dst_p)
             model = self._run_store(dst_path)
             model.title = self.path.name
         model.workflow = self.construct_workflow()
@@ -154,30 +184,52 @@ class RemoteReaderMethod(ReaderMethod):
 
     def run_command(self, dst_path: Path, stdout=None):
         """Run scp/rsync command to move the file from remote to local `dst_path`."""
-        args = remote_to_local(
-            self.protocol,
-            self.to_str(),
-            dst_path,
-            is_wsl=self.wsl,
-            is_dir=self.force_directory,
-            port=self.port,
-        )
+        if isinstance(self.path, Path):
+            args = remote_to_local(
+                self.protocol,
+                self.to_str(),
+                dst_path,
+                is_wsl=self.wsl,
+                is_dir=self.force_directory,
+                port=self.port,
+            )
+        else:
+            raise ValueError(
+                "Cannot run command for multiple paths in RemoteReaderMethod."
+            )
         result = subprocess.run(args, stdout=stdout)
         if result.returncode != 0:
             raise ValueError(f"Failed to run command {args}: {result!r}")
 
 
-class WslReaderMethod(ReaderMethod):
+class WslReaderMethod(PathReaderMethod):
     """Describes that one was read from a WSL source file."""
 
     type: Literal["wsl-reader"] = "wsl-reader"
-    path: Path
-    force_directory: bool = Field(default=False)
+
+    @classmethod
+    def from_str(cls, line: str, force_directory: bool = False) -> "WslReaderMethod":
+        """Construct a WslReaderMethod from a string representation."""
+        if ";" in line:
+            path = [Path(p) for p in line.split(";")]
+        else:
+            path = Path(line)
+        return cls(
+            path=path,
+            force_directory=force_directory,
+        )
 
     def run(self):
         with tempfile.TemporaryDirectory() as tmpdir:
-            dst_path = Path(tmpdir).joinpath(self.path.name)
-            self.run_command(dst_path)
+            if isinstance(self.path, Path):
+                dst_path = Path(tmpdir).joinpath(self.path.name)
+                self.run_command(dst_path)
+            else:
+                dst_path: list[Path] = []
+                for p in self.path:
+                    dst_p = Path(tmpdir).joinpath(p.name)
+                    self.run_command(dst_p)
+                    dst_path.append(dst_p)
             model = self._run_store(dst_path)
             model.title = self.path.name
         model.workflow = self.construct_workflow()
@@ -186,7 +238,14 @@ class WslReaderMethod(ReaderMethod):
     def run_command(self, dst_path: Path, stdout=None):
         """Run cp command to move the file from wsl to local `dst_path`."""
 
-        args = wsl_to_local(self.path, dst_path, is_dir=self.force_directory)
-        result = subprocess.run(args, stdout=stdout)
+        if isinstance(self.path, Path):
+            args = wsl_to_local(
+                self.path.as_posix(), dst_path, is_dir=self.force_directory
+            )
+            result = subprocess.run(args, stdout=stdout)
+        else:
+            raise ValueError(
+                "Cannot run command for multiple paths in WSLReaderMethod."
+            )
         if result.returncode != 0:
             raise ValueError(f"Failed to run command {args}: {result!r}")
