+++ src/himena_builtins/_tests/test_explorer.py
@@ -2,8 +2,10 @@ from pathlib import Path
 from qtpy.QtCore import Qt
 from unittest.mock import MagicMock
 from himena.testing import choose_one_dialog_response
+from himena.workflow import LocalReaderMethod
+from himena_builtins.qt.explorer._base import QBaseRemoteExplorerWidget
 from himena_builtins.qt.explorer._widget import QExplorerWidget
-from himena_builtins.qt.explorer._widget_ssh import QSSHRemoteExplorerWidget, _make_ls_args, _make_get_type_args
+from himena_builtins.qt.explorer._widget_ssh import QSSHRemoteExplorerWidget
 
 from pytestqt.qtbot import QtBot
 
@@ -38,23 +40,53 @@ def test_workspace_widget(qtbot: QtBot, himena_ui, tmpdir):
     # mock.assert_called_once()
     # assert isinstance(mock.call_args[0][0], Path)
 
-def test_ssh_args():
-    assert _make_ls_args("HOST", "PATH") == [
-        "ssh", "-p", "22", "HOST", "ls", "PATH/", "-AF"
-    ]
-    assert _make_get_type_args("HOST", "PATH") == [
-        "ssh", "-p", "22", "HOST", "stat", "PATH", "--format='%F'"
-    ]
-
 def test_ssh_widget(qtbot: QtBot, himena_ui):
     widget = QSSHRemoteExplorerWidget(himena_ui)
     qtbot.add_widget(widget)
     widget.show()
     widget._file_list_widget._make_context_menu()
-    widget._apply_filter("a")
+    widget._file_list_widget._apply_filter("a")
     assert widget._filter_widget.isHidden()
     qtbot.keyClick(widget, Qt.Key.Key_F, Qt.KeyboardModifier.ControlModifier)
     assert widget._filter_widget.isVisible()
     qtbot.keyClick(widget._filter_widget, Qt.Key.Key_Escape)
     assert widget._filter_widget.isHidden()
     widget._file_list_widget._save_items([])
+    widget._make_ls_args("/tmp")
+    widget._make_get_type_args("/tmp")
+
+class QTestRemoteExplorerWidget(QBaseRemoteExplorerWidget):
+    def _make_reader_method(self, path: Path, is_dir: bool) -> LocalReaderMethod:
+        raise LocalReaderMethod(path=path)
+
+    def _set_current_path(self, path: Path):
+        """Set the current path and update the UI accordingly."""
+        self._pwd = path
+
+    def _make_ls_args(self, path: str) -> list[str]:
+        return ["ls", path]
+
+    def _make_get_type_args(self, path: str) -> list[str]:
+        """Make the command to get the type of a file."""
+        return ["stat", path, "--format='%F'"]
+
+    def _make_local_to_remote_args(
+        self, src: Path, dst_remote: str, is_dir: bool = False
+    ) -> list[str]:
+        """Make the command to send a local file to the remote host."""
+        return ["cp", src.as_posix(), dst_remote]
+
+def test_remote_base_widget(qtbot: QtBot, himena_ui):
+    widget = QTestRemoteExplorerWidget(himena_ui)
+    qtbot.add_widget(widget)
+    widget.show()
+    widget._file_list_widget._make_context_menu()
+    widget._file_list_widget._apply_filter("a")
+    assert widget._filter_widget.isHidden()
+    qtbot.keyClick(widget, Qt.Key.Key_F, Qt.KeyboardModifier.ControlModifier)
+    assert widget._filter_widget.isVisible()
+    qtbot.keyClick(widget._filter_widget, Qt.Key.Key_Escape)
+    assert widget._filter_widget.isHidden()
+    widget._file_list_widget._save_items([])
+    widget._make_ls_args("/tmp")
+    widget._make_get_type_args("/tmp")
